requirements.txt
# ─────────────────────────────────────────
# AX Decision Intelligence Engine
# Python 3.11+ required
# ─────────────────────────────────────────

# API Framework
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.1
pydantic-settings==2.3.0

# Async HTTP
httpx==0.27.0
aiohttp==3.9.5
aiofiles==23.2.1

# Scraping & Browser Automation
playwright==1.44.0
beautifulsoup4==4.12.3
lxml==5.2.2
cssselect==1.2.0
selectolax==0.3.21

# NLP & Text Processing
spacy==3.7.5
spacy-langdetect==0.1.2
transformers==4.41.2
torch==2.3.0
nltk==3.8.1
langdetect==1.0.9

# Email Processing
dnspython==2.6.1
disposable-email-domains==0.0.97
validate-email-address==1

# Phone Processing
phonenumbers==8.13.37

# Queue & Cache
celery[redis]==5.4.0
redis==5.0.6
kombu==5.3.7

# Database
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9

# Proxy & Networking
aiohttp-socks==0.8.4
python-socks==2.4.4
fake-useragent==1.5.1

# Monitoring & Logging
structlog==24.2.0
prometheus-client==0.20.0
sentry-sdk[fastapi]==2.3.1
opentelemetry-sdk==1.24.0
opentelemetry-instrumentation-fastapi==0.45b0

# Rate Limiting
slowapi==0.1.9
limits==3.12.0

# Utilities
python-dotenv==1.0.1
tenacity==8.3.0
orjson==3.10.3
arrow==1.3.0
tldextract==5.1.2
url-normalize==1.4.3
extruct==0.17.0
w3lib==2.2.1

# Data Validation
cerberus==1.3.5
jsonschema==4.22.0

# Caching
aiocache==0.12.2

# Testing
pytest==8.2.2
pytest-asyncio==0.23.7
pytest-mock==3.14.0
httpx==0.27.0
factory-boy==3.3.0
